integrate cloudflare/(r2bucket,r2object,r2objects).go into cloudflare/r2.go
