+++ cloudflare/r2object.go
@@ -1,13 +1,1 @@
 package cloudflare
-
-import (
-	"github.com/syumai/workers/cloudflare/r2"
-)
-
-// R2Object represents Cloudflare R2 object.
-// Deprecated: use r2.Object instead.
-type R2Object = r2.Object
-
-// R2HTTPMetadata represents metadata of R2Object.
-// Deprecated: use r2.HTTPMetadata instead.
-type R2HTTPMetadata = r2.HTTPMetadata
