+++ cloudflare/r2objects.go
@@ -1,9 +1,1 @@
 package cloudflare
-
-import (
-	"github.com/syumai/workers/cloudflare/r2"
-)
-
-// R2Objects represents Cloudflare R2 objects.
-// Deprecated: use r2.Objects instead.
-type R2Objects = r2.Objects
